rename DataQueries to DataSources in dashboard
MIMIR-404
