+++ src/main/resources/react4xp/_entries/Dashboard/containers/DataSources/StatisticsQueries.jsx
