+++ src/main/resources/react4xp/_entries/Dashboard/containers/DataSources/slice.es6
