+++ src/main/resources/react4xp/_entries/Dashboard/containers/Statistics/StatisticsLog.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { useDispatch } from 'react-redux'
 import { WebSocketContext } from '../../utils/websocket/WebsocketProvider'
 import { Button, Modal } from 'react-bootstrap'
-import { requestEventLogData } from '../DataQueries/actions'
+import { requestEventLogData } from '../DataSources/actions'
 import { Accordion } from '@statisticsnorway/ssb-component-library'
 import moment from 'moment/min/moment-with-locales'
 import { groupBy } from 'ramda'
